add keyword searching to Read Donations
